requirements.txt
fastapi 
uvicorn 
pymilvus
redis[hiredis]
halo 
sentence-transformers
transformers==4.49.0
wasabi
PyPDF2
Pillow
PyMuPDF
pdf2image
groq
google
google-generativeai
google-ai-generativelanguage